refactor(logging): register ObjectStatus/ObjectLifespan so Log prints names (TASK-59)
`Log(..., lifespan, ...)` used to not compile because ObjectLifespan
wasn't in IsRegisteredEnum, so every call site wrote
`static_cast<int>(lifespan)` and the log printed `Lifespan=2` instead
of `Lifespan=ObjectLifespan::Scene`. Classic fix-at-wrong-layer:
high-level logging concern patched across every caller.

The existing INNO_ENUM macro declares a *new* enum inside Inno::Enum,
which would break every callsite that names ObjectLifespan in the
plain Inno:: namespace. Added INNO_REGISTER_EXTERNAL_ENUM that only
specializes InnoEnumTraits / IsRegisteredEnum / ToString for an enum
already declared elsewhere. Registered ObjectStatus and
ObjectLifespan with it from Object.h.

Stripped the static_cast<int>() hacks at four call sites
(AssetService, GPUBufferResourceServiceImpl, EntityRegistry ×2).
Log now reads e.g.:
  EntityRegistry::CleanUp(Lifespan=ObjectLifespan::Scene) …

Verified: RenderTest 0, Main 3-frame 0, log lines confirm enum-name
output.

Code-AI-Generated-By: Claude Opus 4.7 (1M context)

+++ Source/Engine/Services/AssetService.cpp
@@ -357,7 +357,7 @@ void AssetService::ReleaseAssetsByLifespan(ObjectLifespan lifespan)
 		}
 	}
 
-	Log(Verbose, "AssetService::ReleaseAssetsByLifespan(", static_cast<int>(lifespan),
+	Log(Verbose, "AssetService::ReleaseAssetsByLifespan(", lifespan,
 		") — released meshes=", l_meshReleased, " materials=", l_matReleased, " textures=", l_texReleased);
 }
